all: reorganize section handling from ipython.ts to utility.ts and navigate.ts
